+++ Quill/External/LibTessDotNet/MeshUtils.cs
@@ -38,7 +38,7 @@ using System.Diagnostics;
 namespace Prowl.Quill.External
 {
 
-    using Real = System.Double;
+    using Real = System.Single;
     namespace LibTessDotNet
 {
     internal struct Vec3
