+++ Samples/Common/VectorFont.cs
@@ -11,48 +11,48 @@ namespace Common
     {
         // Character definitions as arrays of line segments
         // Each segment is defined by two points (x1, y1, x2, y2) as a fraction of character cell (0-1)
-        private static readonly Dictionary<char, double[][]> CharacterDefinitions = new Dictionary<char, double[][]> {
+        private static readonly Dictionary<char, float[][]> CharacterDefinitions = new Dictionary<char, float[][]> {
             // Upper case letters
-            ['A'] = [[0, 1, 0.5, 0, 1, 1], [0.2, 0.6, 0.8, 0.6]],
-            ['B'] = [[0, 0, 0, 1, 0.7, 1, 0.9, 0.8, 0.7, 0.5, 0, 0.5], [0, 0, 0.7, 0, 0.9, 0.2, 0.7, 0.5]],
-            ['C'] = [[0.9, 0.2, 0.7, 0, 0.3, 0, 0.1, 0.2, 0.1, 0.8, 0.3, 1, 0.7, 1, 0.9, 0.8]],
-            ['D'] = [[0, 0, 0, 1, 0.6, 1, 0.9, 0.7, 0.9, 0.3, 0.6, 0, 0, 0]],
-            ['E'] = [[0.9, 0, 0, 0, 0, 1, 0.9, 1], [0, 0.5, 0.6, 0.5]],
-            ['F'] = [[0.9, 0, 0, 0, 0, 1], [0, 0.5, 0.6, 0.5]],
-            ['G'] = [[0.9, 0.2, 0.7, 0, 0.3, 0, 0.1, 0.2, 0.1, 0.8, 0.3, 1, 0.7, 1, 0.9, 0.8, 0.9, 0.5, 0.5, 0.5]],
-            ['H'] = [[0, 0, 0, 1], [1, 0, 1, 1], [0, 0.5, 1, 0.5]],
-            ['I'] = [[0.2, 0, 0.8, 0], [0.5, 0, 0.5, 1], [0.2, 1, 0.8, 1]],
-            ['J'] = [[0.8, 0, 0.8, 0.8, 0.6, 1, 0.3, 1, 0.1, 0.8]],
-            ['K'] = [[0, 0, 0, 1], [0, 0.5, 0.8, 0], [0, 0.5, 0.8, 1]],
-            ['L'] = [[0, 0, 0, 1, 0.9, 1]],
-            ['M'] = [[0, 1, 0, 0, 0.5, 0.4, 1, 0, 1, 1]],
+            ['A'] = [[0, 1, 0.5f, 0, 1, 1], [0.2f, 0.6f, 0.8f, 0.6f]],
+            ['B'] = [[0, 0, 0, 1, 0.7f, 1, 0.9f, 0.8f, 0.7f, 0.5f, 0, 0.5f], [0, 0, 0.7f, 0, 0.9f, 0.2f, 0.7f, 0.5f]],
+            ['C'] = [[0.9f, 0.2f, 0.7f, 0, 0.3f, 0, 0.1f, 0.2f, 0.1f, 0.8f, 0.3f, 1, 0.7f, 1, 0.9f, 0.8f]],
+            ['D'] = [[0, 0, 0, 1, 0.6f, 1, 0.9f, 0.7f, 0.9f, 0.3f, 0.6f, 0, 0, 0]],
+            ['E'] = [[0.9f, 0, 0, 0, 0, 1, 0.9f, 1], [0, 0.5f, 0.6f, 0.5f]],
+            ['F'] = [[0.9f, 0, 0, 0, 0, 1], [0, 0.5f, 0.6f, 0.5f]],
+            ['G'] = [[0.9f, 0.2f, 0.7f, 0, 0.3f, 0, 0.1f, 0.2f, 0.1f, 0.8f, 0.3f, 1, 0.7f, 1, 0.9f, 0.8f, 0.9f, 0.5f, 0.5f, 0.5f]],
+            ['H'] = [[0, 0, 0, 1], [1, 0, 1, 1], [0, 0.5f, 1, 0.5f]],
+            ['I'] = [[0.2f, 0, 0.8f, 0], [0.5f, 0, 0.5f, 1], [0.2f, 1, 0.8f, 1]],
+            ['J'] = [[0.8f, 0, 0.8f, 0.8f, 0.6f, 1, 0.3f, 1, 0.1f, 0.8f]],
+            ['K'] = [[0, 0, 0, 1], [0, 0.5f, 0.8f, 0], [0, 0.5f, 0.8f, 1]],
+            ['L'] = [[0, 0, 0, 1, 0.9f, 1]],
+            ['M'] = [[0, 1, 0, 0, 0.5f, 0.4f, 1, 0, 1, 1]],
             ['N'] = [[0, 1, 0, 0, 1, 1, 1, 0]],
-            ['O'] = [[0.3, 0, 0.7, 0, 1, 0.3, 1, 0.7, 0.7, 1, 0.3, 1, 0, 0.7, 0, 0.3, 0.3, 0]],
-            ['P'] = [[0, 1, 0, 0, 0.7, 0, 0.9, 0.2, 0.7, 0.4, 0, 0.4]],
-            ['Q'] = [[0.3, 0, 0.7, 0, 1, 0.3, 1, 0.7, 0.7, 1, 0.3, 1, 0, 0.7, 0, 0.3, 0.3, 0], [0.6, 0.8, 1, 1]],
-            ['R'] = [[0, 1, 0, 0, 0.7, 0, 0.9, 0.2, 0.7, 0.4, 0, 0.4], [0.4, 0.4, 0.9, 1]],
-            ['S'] = [[0.9, 0.2, 0.7, 0, 0.3, 0, 0.1, 0.2, 0.3, 0.4, 0.7, 0.6, 0.9, 0.8, 0.7, 1, 0.3, 1, 0.1, 0.8]],
-            ['T'] = [[0, 0, 1, 0], [0.5, 0, 0.5, 1]],
-            ['U'] = [[0, 0, 0, 0.7, 0.3, 1, 0.7, 1, 1, 0.7, 1, 0]],
-            ['V'] = [[0, 0, 0.5, 1, 1, 0]],
-            ['W'] = [[0, 0, 0.2, 1, 0.5, 0.5, 0.8, 1, 1, 0]],
+            ['O'] = [[0.3f, 0, 0.7f, 0, 1, 0.3f, 1, 0.7f, 0.7f, 1, 0.3f, 1, 0, 0.7f, 0, 0.3f, 0.3f, 0]],
+            ['P'] = [[0, 1, 0, 0, 0.7f, 0, 0.9f, 0.2f, 0.7f, 0.4f, 0, 0.4f]],
+            ['Q'] = [[0.3f, 0, 0.7f, 0, 1, 0.3f, 1, 0.7f, 0.7f, 1, 0.3f, 1, 0, 0.7f, 0, 0.3f, 0.3f, 0], [0.6f, 0.8f, 1, 1]],
+            ['R'] = [[0, 1, 0, 0, 0.7f, 0, 0.9f, 0.2f, 0.7f, 0.4f, 0, 0.4f], [0.4f, 0.4f, 0.9f, 1]],
+            ['S'] = [[0.9f, 0.2f, 0.7f, 0, 0.3f, 0, 0.1f, 0.2f, 0.3f, 0.4f, 0.7f, 0.6f, 0.9f, 0.8f, 0.7f, 1, 0.3f, 1, 0.1f, 0.8f]],
+            ['T'] = [[0, 0, 1, 0], [0.5f, 0, 0.5f, 1]],
+            ['U'] = [[0, 0, 0, 0.7f, 0.3f, 1, 0.7f, 1, 1, 0.7f, 1, 0]],
+            ['V'] = [[0, 0, 0.5f, 1, 1, 0]],
+            ['W'] = [[0, 0, 0.2f, 1, 0.5f, 0.5f, 0.8f, 1, 1, 0]],
             ['X'] = [[0, 0, 1, 1], [0, 1, 1, 0]],
-            ['Y'] = [[0, 0, 0.5, 0.5, 1, 0], [0.5, 0.5, 0.5, 1]],
+            ['Y'] = [[0, 0, 0.5f, 0.5f, 1, 0], [0.5f, 0.5f, 0.5f, 1]],
             ['Z'] = [[0, 0, 1, 0, 0, 1, 1, 1]],
 
             //// Numbers// Numbers
-            ['0'] = [[0.3, 0, 0.7, 0, 1, 0.3, 1, 0.7, 0.7, 1, 0.3, 1, 0, 0.7, 0, 0.3, 0.3, 0], [0, 0.1, 1, 0.9]],
-            ['1'] = [[0.3, 0.2, 0.5, 0, 0.5, 1], [0.3, 1, 0.7, 1]],
-            ['2'] = [[0.1, 0.2, 0.3, 0, 0.7, 0, 0.9, 0.2, 0.9, 0.4, 0.1, 1, 0.9, 1]],
-            ['3'] = [[0.1, 0.2, 0.3, 0, 0.7, 0, 0.9, 0.2, 0.9, 0.4, 0.5, 0.5, 0.9, 0.6, 0.9, 0.8, 0.7, 1, 0.3, 1, 0.1, 0.8]],
-            ['4'] = [[0.7, 0, 0.7, 1], [0, 0.7, 0.9, 0.7], [0.7, 0, 0, 0.7]],
-            ['5'] = [[0.9, 0, 0.1, 0, 0.1, 0.5, 0.7, 0.5, 0.9, 0.7, 0.9, 0.9, 0.7, 1, 0.3, 1, 0.1, 0.9]],
-            ['6'] = [[0.9, 0.1, 0.7, 0, 0.3, 0, 0.1, 0.2, 0.1, 0.8, 0.3, 1, 0.7, 1, 0.9, 0.8, 0.9, 0.6, 0.7, 0.5, 0.3, 0.5, 0.1, 0.6]],
-            ['7'] = [[0.1, 0, 0.9, 0, 0.5, 1], [0.3, 0.5, 0.7, 0.5]],
-            ['8'] = [[0.3, 0, 0.7, 0, 0.9, 0.2, 0.9, 0.3, 0.7, 0.5, 0.3, 0.5, 0.1, 0.3, 0.1, 0.2, 0.3, 0], [0.3, 0.5, 0.1, 0.7, 0.1, 0.8, 0.3, 1, 0.7, 1, 0.9, 0.8, 0.9, 0.7, 0.7, 0.5]],
-            ['9'] = [[0.1, 0.9, 0.3, 1, 0.7, 1, 0.9, 0.8, 0.9, 0.2, 0.7, 0, 0.3, 0, 0.1, 0.2, 0.1, 0.4, 0.3, 0.5, 0.7, 0.5, 0.9, 0.4]],
+            ['0'] = [[0.3f, 0, 0.7f, 0, 1, 0.3f, 1, 0.7f, 0.7f, 1, 0.3f, 1, 0, 0.7f, 0, 0.3f, 0.3f, 0], [0, 0.1f, 1, 0.9f]],
+            ['1'] = [[0.3f, 0.2f, 0.5f, 0, 0.5f, 1], [0.3f, 1, 0.7f, 1]],
+            ['2'] = [[0.1f, 0.2f, 0.3f, 0, 0.7f, 0, 0.9f, 0.2f, 0.9f, 0.4f, 0.1f, 1, 0.9f, 1]],
+            ['3'] = [[0.1f, 0.2f, 0.3f, 0, 0.7f, 0, 0.9f, 0.2f, 0.9f, 0.4f, 0.5f, 0.5f, 0.9f, 0.6f, 0.9f, 0.8f, 0.7f, 1, 0.3f, 1, 0.1f, 0.8f]],
+            ['4'] = [[0.7f, 0, 0.7f, 1], [0, 0.7f, 0.9f, 0.7f], [0.7f, 0, 0, 0.7f]],
+            ['5'] = [[0.9f, 0, 0.1f, 0, 0.1f, 0.5f, 0.7f, 0.5f, 0.9f, 0.7f, 0.9f, 0.9f, 0.7f, 1, 0.3f, 1, 0.1f, 0.9f]],
+            ['6'] = [[0.9f, 0.1f, 0.7f, 0, 0.3f, 0, 0.1f, 0.2f, 0.1f, 0.8f, 0.3f, 1, 0.7f, 1, 0.9f, 0.8f, 0.9f, 0.6f, 0.7f, 0.5f, 0.3f, 0.5f, 0.1f, 0.6f]],
+            ['7'] = [[0.1f, 0, 0.9f, 0, 0.5f, 1], [0.3f, 0.5f, 0.7f, 0.5f]],
+            ['8'] = [[0.3f, 0, 0.7f, 0, 0.9f, 0.2f, 0.9f, 0.3f, 0.7f, 0.5f, 0.3f, 0.5f, 0.1f, 0.3f, 0.1f, 0.2f, 0.3f, 0], [0.3f, 0.5f, 0.1f, 0.7f, 0.1f, 0.8f, 0.3f, 1, 0.7f, 1, 0.9f, 0.8f, 0.9f, 0.7f, 0.7f, 0.5f]],
+            ['9'] = [[0.1f, 0.9f, 0.3f, 1, 0.7f, 1, 0.9f, 0.8f, 0.9f, 0.2f, 0.7f, 0, 0.3f, 0, 0.1f, 0.2f, 0.1f, 0.4f, 0.3f, 0.5f, 0.7f, 0.5f, 0.9f, 0.4f]],
 
-            ['.'] = [[0.5, 1.0, 0.5, 0.8]],
+            ['.'] = [[0.5f, 1.0f, 0.5f, 0.8f]],
             [' '] = []  // Space character has no lines
         };
 
@@ -67,14 +67,14 @@ namespace Common
         /// <param name="color">Color of the text</param>
         /// <param name="lineWidth">Width of the strokes</param>
         /// <param name="spacing">Spacing between characters (0-1)</param>
-        public static void DrawString(Canvas canvas, string text, double x, double y, double height, Color32 color, double lineWidth = 1.0f, double spacing = 0.3f)
+        public static void DrawString(Canvas canvas, string text, float x, float y, float height, Color32 color, float lineWidth = 1.0f, float spacing = 0.3f)
         {
             if (string.IsNullOrEmpty(text))
                 return;
 
-            double charWidth = height * 0.6f;
-            double totalWidth = (charWidth * text.Length) + (spacing * height * (text.Length - 1));
-            double currentX = x;
+            float charWidth = height * 0.6f;
+            float totalWidth = (charWidth * text.Length) + (spacing * height * (text.Length - 1));
+            float currentX = x;
 
             canvas.SaveState();
 
@@ -91,19 +91,19 @@ namespace Common
             canvas.RestoreState();
         }
 
-        public static double MeasureString(string text, double height, double spacing = 0.3) => ((height * 0.6) + (spacing * height))  * (text.Length);
+        public static float MeasureString(string text, float height, float spacing = 0.3f) => ((height * 0.6f) + (spacing * height))  * (text.Length);
 
         /// <summary>
         /// Draws a string using vector lines with the text centered at the given position
         /// </summary>
-        public static void DrawStringCentered(Canvas canvas, string text, double x, double y, double height, Color32 color, double lineWidth = 1.0f, double spacing = 0.2f)
+        public static void DrawStringCentered(Canvas canvas, string text, float x, float y, float height, Color32 color, float lineWidth = 1.0f, float spacing = 0.2f)
         {
             if (string.IsNullOrEmpty(text))
                 return;
 
-            double charWidth = height * 0.6f;
-            double totalWidth = (charWidth * text.Length) + (spacing * height * (text.Length - 1));
-            double startX = x - (totalWidth / 2);
+            float charWidth = height * 0.6f;
+            float totalWidth = (charWidth * text.Length) + (spacing * height * (text.Length - 1));
+            float startX = x - (totalWidth / 2);
 
             DrawString(canvas, text, startX, y, height, color, lineWidth, spacing);
         }
@@ -111,17 +111,17 @@ namespace Common
         /// <summary>
         /// Draws a character using vector lines
         /// </summary>
-        private static void DrawCharacter(Canvas canvas, char c, double x, double y, double height)
+        private static void DrawCharacter(Canvas canvas, char c, float x, float y, float height)
         {
             // Default to space if character not found
-            if (!CharacterDefinitions.TryGetValue(c, out double[][] sections))
+            if (!CharacterDefinitions.TryGetValue(c, out float[][] sections))
                 return;
 
             // Skip if no lines (space)
             if (sections.Length == 0)
                 return;
 
-            double width = height * 0.6f;
+            float width = height * 0.6f;
 
             // Draw each line segment
             for (int i = 0; i < sections.Length; i++)
